Small fixes. pow(-1.0,(int)i) instead of pow(-1.0,i) which is ambiguous to MSVC
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/soc2010/hair@34314 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/material/furmaterial/furmaterial.cpp
@@ -463,7 +463,7 @@ CS_PLUGIN_NAMESPACE_BEGIN(FurMaterial)
 	  for ( size_t j = 0 ; j < hairStrand->guideHairsCount ; j ++ )
 	    hairStrand->controlPoints[i] += hairStrand->guideHairs[j].distance * (
 		  furMaterial->guideHairs.Get(hairStrand->guideHairs[j].index).controlPoints[i] + 
-		  (int)pow(-1,i) * csVector3(0.0f,0,0.0f));
+		  (int)pow(-1.0,(int)i) * csVector3(0.0f,0,0.0f));
 	}
   }
